refactor: flip UserAdministrationService.createUser package-private
EmailUniquenessIntegrationTest was the only cross-package caller of
createUser. Its real subject is the per-tenant uniqueness contract on
UserAdministrationService — a useradmin behaviour test misfiled in the
user/ package. Relocating it to useradmin/ makes it same-package and
removes the only force keeping createUser public.

Part of a broader rule: integration-test fixture state goes through
repositories, not services. Services have production-scenario contracts
(actor required, audit events fire, force-password-change hardcoded)
that don't fit fixture seeding. The test that actually IS the SUT for
the service contract belongs in the service's own package.
